Large commit - reorganizing the dash directory
